requirements.txt
# Core Django and PostgreSQL support
Django==4.2.7
psycopg2-binary==2.9.7
python-decouple==3.8

# Django REST Framework for API development
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7

# Authentication and security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# Database and migrations
django-extensions==3.2.3

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Development and testing
django-debug-toolbar==4.2.0
coverage==7.3.2

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3